demo/requirements.txt
numpy==2.0.0
scikit-learn==1.5.1
pillow==10.4.0
matplotlib==3.9.1
PyGObject==3.48.2
